pyproject.toml
[tool.poetry]
name = "minimal_vertex_pipeline"
version = "0.1.0"
description = ""
authors = ["David Aronchick <aronchick@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
kfp = "^1.8.12"
dill = "^0.3.5"
google-cloud-aiplatform = "^1.14.0"
click = "^8.1.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
